+++ src/ExtensionMethod.psm1
@@ -1,18 +1,8 @@
 ﻿#ExtensionMethod.psm1
 
-#Todo :
-# GetValue([System.Management.Automation.Language.NamedAttributeArgumentAst] $attrAst, [System.Management.Automation.Language.ExpressionAst&amp;] $argumentAst)
-# les parametres Ref et Out  sont poste fixé par & 
-#    System.Management.Automation.Language.ExpressionAst&
-#infos méthode
-#   ParameterType    : System.Management.Automation.Language.ExpressionAst&
-#   Name             : argumentAst
-#   Attributes       : Out
-#   ...
-
-# parametertype.isbyref -> true -> c'est un 'Out parametre'
-#Type realType = pInfo.ParameterType.GetElementType();
-#cf https://stackoverflow.com/questions/738277/net-reflection-how-to-get-real-type-from-out-parameterinfo
+#todo : [Linq.Enumerable]::Sum([int[]]($this.Compteur))
+#si le premier paramètre est de type IEnumerable un cast peut suffire
+# public static int Sum(	this IEnumerable<int> source)
 
 #Fonctions d'aide à la création de fichier ps1xml dédié aux
 # méthodes d'extension contenues dans un fichier assembly dotnet.
@@ -93,6 +83,29 @@ function Test-ClassExtensionMethod([System.Type] $Type)
     }).count -gt 0
 }#Test-ClassExtensionMethod
 
+function AddMembers{
+ #Add to a MethodInfo the following members : ParameterCount,CountOptional,isContainsParams
+ #  ParameterCount= Number of parameter
+ #  CountOptional= Number of optional parameter
+ #  isContainsParams= $true when a parameter is 'params'
+  param($CurrentMethod)
+
+ $CountOptional=$Count=0
+ $isContainsParams=$false
+ foreach ($CurrentParameter in $CurrentMethod.GetParameters())
+ {
+   $Count++
+   if ($isContainsParams -eq $false)
+   { $isContainsParams=$CurrentParameter.GetCustomAttributes([System.ParamArrayAttribute],$false).Count -gt 0 }
+   if ($CurrentParameter.isOptional)
+   { $CountOptional++ }
+ }
+ Add-member -inputobject $CurrentMethod -membertype NoteProperty -name ParameterCount -value $Count -passthru|
+   Add-member -membertype NoteProperty -name CountOptional -value $CountOptional -passthru|
+   Add-member -membertype NoteProperty -name isContainsParams -value $isContainsParams
+ write-output $CurrentMethod
+}#AddMembers
+
 function Get-ExtensionMethodInfo{
   #Renvoi, à partir d'informations d'une méthode d'extension, un objet DictionaryEntry dont :
   #       la clé est le type du premier paramètre, déclaré dans la signature de la méthode (le type peut être une type imbriqué),
@@ -134,13 +147,14 @@ function Get-ExtensionMethodInfo{
            )
          ))
       {
-         #Pas d'exclusion demandée ou le type n'est pas générique.
+        $MethodInfo= AddMembers -CurrentMethod $_
+        #Pas d'exclusion demandée ou le type n'est pas générique.
          #On renvoi une paire clé/valeur avec le type du premier paramètre et l'objet méthode
         new-object System.Collections.DictionaryEntry(
                                          #la clé est le nom du type du premier paramètre de l'objet méthode
                                        ($Parameter.ParameterType.ToString()),
                                          #la valeur est l'objet méthode
-                                        $_
+                                        $MethodInfo
                                       )
       } #if ExcludeGeneric
     } #if ExcludeInterface
@@ -160,7 +174,7 @@ Function New-HashTable {
 param(
      [Parameter(Position=0, Mandatory=$true,ValueFromPipeline = $true)]
     $InputObject,
-    
+
      [Parameter(Mandatory=$true)]
     $Key,
     $Value,
@@ -211,32 +225,34 @@ function Format-TableExtensionMethod{
      #Contient des méthodes d'extension reroupées par nom de type.
     [System.Collections.HashTable] $Hashtable
  )
-
-$Hashtable.GetEnumerator()|
- Foreach-Object {
-   $InstanceType=$_.Key;$_.Value|
-    Sort-Object $_.ParameterCount|
-    Add-Member NoteProperty InstanceType $InstanceType -pass
- }|
- Format-Table @{Label="Method";Expression={ $_.ToString()}} -GroupBy InstanceType
+process {
+  $Hashtable.GetEnumerator()|
+  Foreach-Object {
+    $InstanceType=$_.Key;$_.Value|
+      Sort-Object $_.ParameterCount|
+      Add-Member NoteProperty InstanceType $InstanceType -pass
+  }|
+  Format-Table @{Label="Method";Expression={ $_.ToString()}} -GroupBy InstanceType
+}
 }#Format-TableExtensionMethod
 
 Function Get-ParameterComment {
-# need dotNET 4.5 -> ps v4 and > 
+# need dotNET 4.5 -> ps v4 and >
+#todo vérifier si toutes les signatures d'une method surchargée sont précisées
  param( $Method )
   $MethodSignature= foreach($Parameter in $Method.GetParameters()) {
-    
+
     $ParameterName=$Parameter.Name
     Write-Debug "$ParameterName"
     if ($Parameter.ParameterType.IsByRef)
-    { 
+    {
       Write-Debug " byRef"
       #[ref] powershell is for 'out' and 'ref' c# parameter modifier
       $ParameterStatement="[ref] [$($Parameter.ParameterType.GetElementType())]"
     }
     else
     {
-      $ParameterStatement="[$($Parameter.ParameterType)]" #default  
+      $ParameterStatement="[$($Parameter.ParameterType)]" #default
       if ($Parameter.IsOptional)
       {
         Write-Debug " optional"
@@ -246,7 +262,7 @@ Function Get-ParameterComment {
            Write-Debug "  hasDefault"
            $isPtr=($ParameterType.FullName -eq 'System.IntPtr') -or ($ParameterType.FullName -eq 'System.UIntPtr')
            if ($null -eq $Parameter.DefaultValue)
-           { 
+           {
              $ParameterName +="=`$null"
              if ($isPtr)
              {$ParameterName +=" OR new OR default" } #impossible to determine one of the three cases ?
@@ -254,33 +270,33 @@ Function Get-ParameterComment {
            else
            {
              Write-Debug "   NotNull value"
-             if (($Parameter.RawDefaultValue -is [string]) -or (($Parameter.RawDefaultValue -is [char]) ))  
+             if (($Parameter.RawDefaultValue -is [string]) -or (($Parameter.RawDefaultValue -is [char]) ))
              {
                Write-Debug "     String or Char"
-               $ParameterName +="='$($Parameter.RawDefaultValue)'" 
+               $ParameterName +="='$($Parameter.RawDefaultValue)'"
              }
              elseif ($ParameterType.isEnum)
-             { 
+             {
                Write-Debug "     ENUM"
-               $ParameterName +="='$([system.Enum]::Parse($Parameter.ParameterType,$Parameter.RawDefaultValue))'" 
+               $ParameterName +="='$([system.Enum]::Parse($Parameter.ParameterType,$Parameter.RawDefaultValue))'"
              }
              elseif ($Parameter.RawDefaultValue -is [boolean])
-             { 
+             {
                Write-Debug "     bool"
-               $ParameterName +="=`$$($Parameter.RawDefaultValue)" 
+               $ParameterName +="=`$$($Parameter.RawDefaultValue)"
              }
              elseif (($ParameterType.isPrimitive -and (-not $isPtr)) -or ($Parameter.RawDefaultValue -is [Decimal]) -or ($Parameter.RawDefaultValue -is [Single]))
-             { 
+             {
                Write-Debug "     Numeric"
-               $ParameterName +="=$($Parameter.DefaultValue.ToString([System.Globalization.CultureInfo]::InvariantCulture))" 
+               $ParameterName +="=$($Parameter.DefaultValue.ToString([System.Globalization.CultureInfo]::InvariantCulture))"
              }
              #else TODO ? isArray Or IEnumerable
-           }  
+           }
         }#hasDefaultValue
         else
-        { 
+        {
           Write-Debug "     Default"
-          $ParameterName +="=new OR default" 
+          $ParameterName +="=new OR default"
         }
       }#isOptionnal
     }
@@ -303,6 +319,41 @@ Function New-ExtensionMethodType{
    # DictionaryEntry contenant les méthodes d'extension d'une classe,
  [System.Collections.DictionaryEntry] $Entry
  )
+begin {
+   function AddAllParameters{
+     param($Count)
+    foreach ($Number in 1..($Count-1))
+    {
+      Write-output ",`$args[$($Number-1)]"
+    }
+   }
+
+  function AddSwitchClauseForMethodWithParams{
+    param($ScriptBuilder,$MaxSignatureWithParamsKeyWord)
+
+    if ($null -ne $MaxSignatureWithParamsKeyWord)
+    {
+        #Add last signature with Params
+      $MaxSignatureWithoutParams=$SortedTemp|Where-Object {-not $_.isContainsParams} |Select-Object -first 1
+      $Max=[Math]::Max($MaxSignatureWithoutParams.ParameterCount,$MaxSignatureWithParamsKeyWord.ParameterCount)
+
+      $Script.Append( ("`t`t`t`t $(Get-ParameterComment -Method $MaxSignatureWithParamsKeyWord)")) >$null
+      $arguments=AddAllParameters -Count ($Max-1)
+      #todo params peut ne pas être renseigné dans l'appel, dans ce cas
+      # on a {1} {{ [Object[]]`$Params=@(`$this {0})" -f "$Arguments",($Max-2))
+      #Si le cas ($Max-2) Args.Count n'existe pas
+      $Script.AppendLine(("`t`t {{`$_ -gt {1}}} {{ [Object[]]`$Params=@(`$this {0},@(`$args[{1}..`$args.count-1]))" -f "$Arguments",($Max-2)) ) >$null
+      $Script.AppendLine(('                  [{0}]::{1}.Invoke($Params)' -f $MaxSignatureWithParamsKeyWord.Declaringtype,$MethodName) ) >$null
+      $Script.AppendLine('                }') >$null
+    }
+  # (voirs les liens ) https://stackoverflow.com/questions/6484651/calling-a-function-using-reflection-that-has-a-params-parameter-methodbase
+  #                    https://stackoverflow.com/questions/23660137/c-sharp-reflective-method-invocation-with-arbitrary-number-of-parameters?noredirect=1&lq=1
+  #                    https://stackoverflow.com/questions/22235490/methodinfo-invoke-throws-exception-for-variable-number-of-arguments?noredirect=1&lq=1
+  #                    https://stackoverflow.com/questions/16777547/invoke-a-method-using-reflection-with-the-params-keyword-without-arguments?noredirect=1&lq=1
+  #
+  #                    https://stackoverflow.com/questions/35404295/invoking-generic-method-with-params-parameter-through-reflection
+  }
+}#begin
 
 #Pour chaque type, on crée autant de balise <ScriptMethod> que de méthodes recensées.
 #Comme chaque méthode peut être surchargée,  on doit considérer le type de la surcharge,
@@ -328,39 +379,54 @@ Function New-ExtensionMethodType{
 #Si le nombre de paramètres correspond, mais pas leurs type alors le shell déclenchera une exception.
 
  process {
-    $TypeName=$Entry.Key
-    $Methods=$Entry.Value
-    Write-Verbose "Type '$TypeName'"
-    New-Type -Name $TypeName -Members {
+   $TypeName=$Entry.Key
+   $MethodsInfo=$Entry.Value
+   Write-Verbose "Type '$TypeName'"
 
-    $Methods| Group-Object Name |
-      #Pour chaque groupe de méthodes du type courant
-      #Utilise une Here-String imbriquées
-    Foreach-Object {
-      Write-Verbose "`tMethod '$($_.Name)'"
-        #ToString() peut provoquer des appels récursifs fatal, par exemple sur System.Object
-      if ( $_.Name -eq 'ToString')
-      { Write-Warning "Excluded method : $TypeName.ToString()" }
+   New-Type -Name $TypeName -Members {
+    foreach ($GroupMethod in $MethodsInfo| Group-Object Name)
+    {
+      $MethodName=$GroupMethod.Name
+      Write-Verbose "`tMethod '$MethodName'"
+      if ( $MethodName -eq 'ToString')
+      {
+         #ToString () can cause fatal recursive calls, for example on System.Object
+        Write-Warning "Excluded method : $TypeName.ToString()"
+      }
       else
       {
-        #begin switch
         $Script=New-Object System.Text.StringBuilder " switch (`$args.Count) {`r`n"
-        $_.group|
-            #Pour faciliter le tri on ajoute un membre contenant le nombre de paramètre de la méthode.
-         Add-member ScriptProperty ParameterCount -value {$this.GetParameters().Count} -Force -PassThru|
-            #On ne crée qu'une seule ligne pour les méthodes surchargées utilisant un nombre identique de paramètres.
-            #Exemple :
-            #  public static SubStringFrom From(this string s, int start);
-            #  public static SubStringFrom From(this string s, string start);
-            #Dans ce cas leurs types est différent, ce qui ne pose pas de pb car PowerShell est non typé.
-            #note : pas de gestion nécessaire pour  Ref et Out -> Compiler Error CS1101
-         Sort-Object ParameterCount -unique|
-         Foreach-Object {
-            $Count=$_.ParameterCount
-                #On soustrait 1 à $Count pour créer un décalage :
-                #    0=$this                    -> $Objet.Method()
-                #    1=$this,$args[0]           -> $Objet.Method($Param1)
-                #    2=$this,$args[0],$args[1]  -> $Objet.Method($Param1,$Param2)
+
+         #contient les numéros de section du switch associé à une méthode.
+         #Le numéros de section est le nombre de paramètres de chaque signature
+        $SwitchSections= [System.Collections.Generic.HashSet[String]]::new()
+
+          #Only one line is created for overloaded methods using the same number of parameters.
+          #Example :
+          #  public static SubStringFrom From(this string s, int start);
+          #  public static SubStringFrom From(this string s, string start);
+          #In this case their types are different, which is not a problem because PowerShell is untyped.
+          #note: The parameter modifier 'ref' cannot be used with 'this' -> Compiler Error CS1101
+
+        $GroupMethodSignatures=$GroupMethod.group |Group-Object ParameterCount
+        $SortedTemp=$GroupMethod.group|Sort-Object parametercount -Descending
+        $MaxSignatureWithParams=$SortedTemp|Where-Object isContainsParams |Select-Object -first 1
+
+        foreach ($Method in $GroupMethodSignatures)
+        {
+            if ($null -ne $MaxSignatureWithParams)
+            {
+              #The method declaring Params is added last, we do not duplicate his statement
+              $CurrentSignatureWithParams=$Method.Group|Where-Object isContainsParams |Select-Object -first 1
+              if (($null -ne $CurrentSignatureWithParams) -and $CurrentSignatureWithParams.Equals($MaxSignatureWithParams))
+              { continue }
+            }
+            [int]$ParameterCount=$Method.Name
+            $SwitchSections.Add($ParameterCount)>$null
+                #On soustrait 1 à $ParameterCount pour créer un décalage :
+                #    0=$this                    -> $Objet.Method()                 -> [Type]::Method($this)
+                #    1=$this,$args[0]           -> $Objet.Method($Param1)          -> [Type]::Method($this,$args[0])
+                #    2=$this,$args[0],$args[1]  -> $Objet.Method($Param1,$Param2)  -> [Type]::Method($this,$args[0],$args[1])
                 #
                 # Le décalage sur le nombre de paramètres est dû au fait que l'on doit prendre en charge
                 # le modificateur C# this, qui est égal à $this dans un script de méthode PowerShell, celui-ci
@@ -368,32 +434,141 @@ Function New-ExtensionMethodType{
                 #
                 #On construit plusieurs lignes respectant la syntaxe d'appel d'une méthode d'extension( méthode statique) :
                 #  nb_argument  { [TypeName]::MethodName($this, 0..n arguments) }
-            #current case
-            $Script.Append( ("`t`t`t`t $(Get-ParameterComment -Method $_)")) >$null
-            $Script.Append( ("`t`t {0} {{ [{1}]::{2}(`$this" -F ($Count-1), $_.Declaringtype,$_.Name)) >$null
-            if ($Count -gt 1)
+                #
+                #Note : Si un paramétre est ByRef c'est l'appelant qui le déclare [ref] sur la variable utilisée et PS le propage à la méthode d'extension
+
+             #switch clause for ($ParameterCount-1) $Args
+            foreach ( $MethodToComment in $Method.Group)
+            {
+               #Add all method signatures with the same number of parameters
+               #todo :     s'il existe pas dans le groupe une méthode avec le même nb de param on ne l'ajoute pas celle avec params
+               #           s'il n'existe pas une méthode avec param +1
+              $Script.Append( ("`t`t`t`t $(Get-ParameterComment -Method $MethodToComment)")) >$null
+            }
+            $Script.Append( ("`t`t {0} {{ [{1}]::{2}(`$this" -F ($ParameterCount-1), $Method.group[0].Declaringtype,$MethodName)) >$null
+            if ($ParameterCount -gt 1)
             {
                 $ofs=''
-                $arguments=1..($Count-1)|Foreach-Object { ",`$args[$($_-1)]" }
+                $arguments= AddAllParameters -Count $ParameterCount
                 $Script.Append("$arguments") >$null
             }
             #close method call
             $Script.Append(") }`r`n`r`n") >$null
-         }#foreach
-         #default
-         $Script.Append( ("`t`t default {{ throw `"No overload for '{0}' takes the specified number of parameters.`" }}" -F $_.Name) ) >$null
-         #end switch
-         $Script.Append("`r`n }") >$null
+        }#foreach $Method
 
-         Write-Debug "`tWrite ScriptMethod '$($_.Name)'"
-         Write-debug "`tscript: $Script"
-         New-ScriptMethod -Name $_.Name -Script $Script
+        AddSwitchClauseForMethodWithParams -ScriptBuilder $Script -MaxSignatureWithParamsKeyWord $MaxSignatureWithParams
+
+         #Add default switch clause
+        $Script.Appendline( ("`t`t default {{ throw `"No overload for '{0}' takes the specified number of parameters.`" }}" -F $MethodName) ) >$null
+
+          #Add end of switch
+        $Script.Appendline('   }') >$null
+
+        Write-Debug "`tWrite ScriptMethod '$($MethodName)'"
+        Write-debug "`tscript: $Script"
+        New-ScriptMethod -Name $MethodName -Script $Script
       } #else
-     } #foreach $Methods
-    } #New-Type
+     } #foreach $GroupMethods
+   } #New-Type
  } #process
 } #New-ExtensionMethodType
 
+<#
+todo scénario de construction -> on doit ajouter des cas dans le switch selon la déclaration de méthode
+ traite celles sans option ni params
+ traite celles avec option ni params
+ traite celles sans option et avec params
+ traite celles avec option et avec params ->Error : PS ne gére pas le cas où l'optionnel est absent et params 'vide'
+
+
+    public static string Method1(this string S, bool includeBoundary=true){  # ici on doit ajouter une signature qui n'existe pas dans la liste des méthodes
+                                                                             method1(1 arg)
+
+    pour 2 param :
+     existe-t-il une signature de méthode ayant 1 paramètre ?
+       oui suite
+       non ajoute
+     existe-t-il une signature de méthode ayant 2 paramètres ?
+       oui suite
+       non ajoute
+
+    public static string Method2(this string S, int end)
+    public static string Method2(this string S, params object[] parameters){  # ici on doit ajouter une signature qui n'existe pas dans la liste des méthodes
+                                                                             method1(1 arg)
+                                                                             method1(1 arg, 2 et >)
+    public static string To(this string S, double end, bool includeBoundary){
+
+    public static string Method2(this string S, int end, params object[] parameters){  # todo ici le switch du premier dépend du nb de param des méthodes suivantes
+
+    public static string From(this string S){
+    public static string From(this string S, bool includeBoundary=true){ #ici sans l'optionnel, le compilo appel -> From(this string S)
+                                                                         #ici on n'ajoute pas de signature, car elle existe déjà dans la liste des méthodes
+
+    public static string From(this string S){
+    public static string From(this string S, int end){
+    public static string From(this string S, bool includeBoundary=true){  #ici on n'ajoute pas la signature, car elle existe déjà
+                                                                          #todo c'est le type qui détermine la méthode appeler ?
+
+    public static string To(this string S){
+    public static string To(this string S, bool includeBoundary=true){ #todo on doit ordonner le traitemnt des signatures
+    public static string To(this string S, int end){
+    public static string To(this string S, string end){
+    public static string To(this string S, int end, bool includeBoundary=true){
+    public static string To(this string S, double end, bool includeBoundary=true){
+    public static string To(this string S, double end, bool includeBoundary=true, int count=2){
+
+
+une seule ayant un param, on traite
+  System.String To(System.String)
+
+ $CountArgCreated.Add(1)
+
+deux ayant 2 params, on ne crée qu'une seule entrée pour 2
+  System.String To(System.String, Int32)
+  System.String To(System.String, System.String)
+
+ $CountArgCreated.Add(2)
+
+une ayant 3 params dont un optionnel,on crée une seule entrée pour 3 params
+      si (3         -        1) est déjà déclaré on passe if -not ($CountArgCreated.contains( (3-1) ) {create '2 {call [Class]::Method($this,$args[1],$args[2])}
+      si (3         -        1) n'est déjà déclaré on crée une seule entrée pour 2 params
+
+  $CountArgCreated.Add(3)
+System.String To(System.String, Int32, Boolean) #optional
+
+une ayant 4 params dont 2 optionnels, on crée une seule entrée pour 4 params
+      si (4         -        2) est déjà déclaré on passe
+      si (4         -        1) n'est déjà déclaré on crée une seule entrée pour 2 params
+System.String To(System.String, Int32, Boolean,int) #optionals
+
+$CountArgCreated.Add(4)
+
+
+
+
+                pour les param optionnel On ajoute des signatures dans la hashtable ?
+
+                $OptionnalParameter= $Method.GetParameters()|where isOptionnal ( ne peuvent être byref en même temps
+
+                 2 param mais un optionnnel ( ils sont à la fin)
+                switch ($args.Count) {
+                  # Func1([string] $S, [bool] $isvalid=$True)
+       --->   ajouter  (count-nBParamOption) NUMBER { [BasicTest]::Func1($this) }
+                        (2-1) -> une ligne avec le premier seulement
+                 1 { [BasicTest]::Func1($this,$args[0]) }
+
+                 default { throw "No overload for 'Func1' takes the specified number of parameters." }
+               }
+
+               pour
+                # Add([string] $S, [int] $a=3, [int] $b=5)
+                (3-1) -> deux lignes: première avec this seulement
+                                      seconde avec this +le premier
+                2 { [BasicTest]::Add($this,$args[0],$args[1]) }
+#>
+
+
+#todo construction des blocs a revoir
 function New-ExtendedTypeData {
   #Create an extension file (ETS) containing wrappers of extension methods
  [CmdletBinding(DefaultParameterSetName="Path",SupportsShouldProcess = $true)]
@@ -433,8 +608,8 @@ function New-ExtendedTypeData {
                                                           Justification="WriteDatas create files.The caller declare ShouldProcess.")]
       param (
         [string] $FileName,
-        [string] $Datas,
-        [switch] $isLiteral
+        [string] $TypeData,
+        [switch] $IsLiteral
       )
 
        #Update-TypeData do not manage [] or `[`] for the name of the file : Byte[].ps1xml
@@ -456,9 +631,9 @@ function New-ExtendedTypeData {
        {
           #Only one confirmation
          if ($isLiteral)
-         { Set-Content -Value $Datas -LiteralPath $FileName -Encoding UTF8 -Confirm:$false}
+         { Set-Content -Value $TypeData -LiteralPath $FileName -Encoding UTF8 -Confirm:$false}
          else
-         { Set-Content -Value $Datas -Path $FileName -Encoding UTF8 -Confirm:$false}
+         { Set-Content -Value $TypeData -Path $FileName -Encoding UTF8 -Confirm:$false}
        }
       }#ShouldProcess
     }#WriteDatas
@@ -486,7 +661,7 @@ function New-ExtendedTypeData {
       $EtsDatas=New-TypeData -PreContent '<?xml version="1.0" encoding="utf-8"?>' -Types {
          $ScriptMethods
        }
-      WriteDatas -FileName $FileInfo -Datas $EtsDatas -isLiteral:$isLiteral
+      WriteDatas -FileName $FileInfo -TypeData $EtsDatas -isLiteral:$isLiteral
     }
     else
     {
@@ -499,7 +674,7 @@ function New-ExtendedTypeData {
           $EtsDatas= New-TypeData -PreContent '<?xml version="1.0" encoding="utf-8"?>' -Types {
              $ScriptMethod
            }
-          WriteDatas -FileName $LiteralPath -Datas $EtsDatas -isLiteral
+          WriteDatas -FileName $LiteralPath -TypeData $EtsDatas -isLiteral
        }#foreach
     }#else
  }#end